hotornot/requirements.txt
argcomplete==1.5.1
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.1
bleach==2.0.0
blessings==1.6
bpython==0.16
certifi==2017.4.17
cffi==1.7.0
chardet==3.0.3
configparser==3.5.0
cryptography==1.4
cssmin==0.2.0
curtsies==0.2.11
decorator==4.0.9
Django==1.8
django-admin-bootstrapped==2.5.7
django-angular==0.8.4
django-appconf==1.0.1
django-bootstrap-form==3.2.1
django-common-helpers==0.8.0
django-compressor==2.0
django-cron==0.4.6
django-crontab==0.7.1
django-geoip==0.5.2
django-grappelli==2.8.1
django-haystack==2.4.1
django-helpdesk==0.1.17
django-htmlmin==0.9.1
django-markdown-deux==1.0.5
django-mptt==0.8.5
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
EbookLib==0.15
email-reply-parser==0.3.0
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
gdata==2.0.18
google-api-python-client==1.5.2
greenlet==0.4.12
gunicorn==19.4.5
html5lib==0.999999999
httpie==0.9.9
httplib2==0.9.2
idna==2.5
ipaddress==1.0.16
ipykernel==4.6.1
ipython==4.1.1
ipython-genutils==0.1.0
ipywidgets==6.0.0
Jinja2==2.9.6
jsmin==2.2.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
lxml==3.6.4
markdown2==2.3.1
MarkupSafe==1.0
meld3==1.0.2
mistune==0.7.3
mysqlclient==1.3.7
nbconvert==5.1.1
nbformat==4.3.0
ndg-httpsclient==0.4.2
nose==1.3.7
notebook==5.0.0
numpy==1.10.4
oauth2==1.9.0.post1
oauth2client==1.5.0
oauthlib==1.0.3
pandas==0.18.1
pandocfilters==1.4.1
path.py==8.1.2
paypalrestsdk==1.11.7
pdfminer==20140328
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.1
poster==0.8.1
praw==4.0.0b21
prawcore==0.2.1
prompt-toolkit==1.0.14
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.5.3
pyOpenSSL==16.0.0
pyperclip==1.5.32
pysolr==3.4.0
python-dateutil==2.5.3
python-docx==0.8.6
python-linkedin==4.1
python-memcached==1.57
python-openid==2.2.5
python-pptx==0.6.1
pytz==2016.4
pyzmq==16.0.2
qtconsole==4.3.0
rcssmin==1.0.6
recaptcha-client==1.0.6
requests==2.17.3
requests-oauthlib==0.6.1
rjsmin==1.0.12
rsa==3.4.2
selenium==3.7.0
simplegeneric==0.8.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.3.0
SpeechRecognition==3.4.6
supervisor==3.2.1
terminado==0.6
testpath==0.3
textract==1.4.0
tornado==4.5.1
traitlets==4.3.2
update-checker==0.12
uritemplate==0.6
urllib3==1.21.1
uWSGI==2.0.15
wcwidth==0.1.7
webencodings==0.5
widgetsnbextension==2.0.0
xlrd==0.9.4
XlsxWriter==0.9.2
xlwt==1.0.0
